+++ package.json
@@ -12,7 +12,9 @@
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
-    "react-native": "0.72.6"
+    "react-native": "0.72.6",
+    "react-native-paper": "^5.12.1",
+    "react-native-safe-area-context": "^4.8.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
